requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
astroid==2.11.7
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
black==22.6.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
dill==0.3.5.1
discord.py==2.0.0
frozenlist==1.3.1
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
lxml==4.9.1
mccabe==0.7.0
multidict==6.0.2
mypy-extensions==0.4.3
pathspec==0.9.0
pep8==1.7.1
platformdirs==2.5.2
pylint==2.14.5
python-dotenv==0.20.0
requests==2.28.1
soupsieve==2.3.2.post1
tomli==2.0.1
tomlkit==0.11.4
urllib3==1.26.12
wrapt==1.14.1
yarl==1.8.1
